+++ astrobot/__init__.py
@@ -1,3 +1,3 @@
 """all-encompassing god-level bot for mochjicord"""
 
-__version__ = "22.2.4.patch1"
+__version__ = "22.4.1"